+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Changed
   compiling them directly.
 - Change the CommandModule class' optional unload method to invoke top-level
   unload functions like the Natlink loader does.
+- Make the *Integer*, *IntegerRef* and *ShortIntegerRef* elements treat
+  their optional *max* argument as an inclusive upper bound.
 - Make the logging output of Dragonfly's CLI commands more sane.
 - Make some optimizations to the Natlink engine.
 - Rename the engines.backend_sphinx.misc module to config.